Create AGP or PCI Card Detection
In this modified script, the lspci command is used with the -i option to search for a case-insensitive string that matches "agp". If the string is found, the script sets the agp variable to true and prints a message that an AGP video card is detected. Otherwise, the script sets the agp variable to false and prints a message that a PCI video card is detected.

After starting the OpenGL diagnostic test, the script uses the agp variable to perform actions specific to AGP or PCI cards, such as modifying system settings or running additional tests. Note that the specific actions you take may depend on the hardware and software configuration of your system.