Create Detect Specific Card
This script is a further expansion on the previous scripts. This one attempts to determine the correct card type and only load the correct driver. The $card command in the modified script extracts the name of the graphics card from the output of the lspci command, but the name it extracts may not always match the name used by the driver installation packages. So it's possible that the script may not be able to find and install the correct driver using the extracted name alone.