Create Ubuntu 4.10 Script
This should detect the card on a Linux distro that is specific to the AGP Video Card era. Written for Ubuntu 4.10.

To start this script on system startup, you can add a call to the script in the system's initialization scripts. For example, you can add the following line to the /etc/rc.local file:

/path/to/script.sh &

Make sure to replace /path/to/script.sh with the actual path to the script file. The & at the end of the line runs the script in the background, so that the initialization process can continue.